requirements.txt
flask
flask-restful
flask-sqlalchemy
pyjwt
